refactor(iota-config): move filesystem config helpers into persisted_config module
Address review feedback: extract the node-only Config / PersistedConfig
traits and the filesystem path helpers (iota_config_dir,
genesis_blob_exists, validator_config_file, ssfn_config_file) into a
dedicated persisted_config module gated by a single cfg attribute,
instead of annotating each item individually.
